+++ client/components/Bubble.jsx
