--- client/components/game.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Start  = props => {
-
-    <button>Start</button>
-}
-
-export default Start
-
