+++ client/index.js
@@ -4,6 +4,7 @@ var ReactDOM = require('react-dom')
 import App from './App'
 
 
+
 var placeToMount = document.getElementById('root')
 
 ReactDOM.render(<App width={window.innerWidth} height={window.innerHeight} />, placeToMount)
