+++ public/index.html
@@ -1,6 +1,8 @@
 <html>
+  <head>
+    <link rel="stylesheet" type="text/css" href="./styles.css">
+  </head>
   <body>
-    <h1> Hello World </h1>
     <div id='root'></div>
 
     <script src='bundle.js'></script>
