+++ public/styles.css
@@ -1,1 +1,12 @@
-.
+.startButton {
+    background-color: #add8e6;
+    border: none;
+    color: white;
+    padding: 20px;
+    text-align: center;
+    text-decoration: none;
+    display: inline-block;
+    font-size: 20px;
+    margin: 4px 2px;
+    border-radius: 60%;
+  }
